Making progress with the tutorial. Now able to run the store and an external catalog on different SCA nodes.
git-svn-id: https://svn.apache.org/repos/asf/incubator/tuscany/java/sca@588297 13f79535-47bb-0310-9956-ffa450edef68

+++ tutorial/store/src/main/java/services/CurrencyConverter.java
@@ -23,7 +23,7 @@ import org.osoa.sca.annotations.Remotable;
 
 @Remotable
 public interface CurrencyConverter {
-    public float getConversion(String fromCurrenycCode, String toCurrencyCode, float amount);
+    public double getConversion(String fromCurrenycCode, String toCurrencyCode, double amount);
 
     public String getCurrencySymbol(String currencyCode);
 }
